+++ packages/react-magma-dom/src/index.ts
@@ -73,6 +73,7 @@ export {
   DropdownButton,
   DropdownButtonProps,
 } from './components/Dropdown/DropdownButton';
+export { Form, FormProps } from './components/Form';
 export { FormGroup, FormGroupProps } from './components/FormGroup';
 export { Heading, HeadingProps } from './components/Heading';
 export { HideAtBreakpoint } from './components/HideAtBreakpoint';
@@ -138,6 +139,7 @@ export { Toggle, ToggleTextPosition, ToggleProps } from './components/Toggle';
 export {
   Tooltip,
   EnumTooltipPosition,
+  TooltipPosition,
   TooltipProps,
 } from './components/Tooltip';
 export { Textarea, TextareaProps } from './components/Textarea';